dataprocessing/requirements.txt
langchain>=0.1.0
networkx>=3.1
pandas>=2.0.0
numpy>=1.24.0
spacy>=3.7.0
python-docx>=1.0.0
PyPDF2>=3.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pinecone>=2.2.4 